Move Message and `form` into `internal/form` package
Fixes #297
