+++ internal/api/message.go
@@ -1,0 +1,6 @@
+package api
+
+// A message is a request/response type, used in api.Stub
+type Message interface {
+	Validate() error
+}
